Remove go get from the Readme and favor dep ensure.
This ammends existing instructions for dep usage to something that is
easier to manage as there is no need for the `dep ensure -add` step.

If you do `dep ensure -add` then you also need to make sure you manually
remove the entries from Gopkg.toml when you stop using a library.
Using the dep feature where it introspects your includes is a less
error prone way to deal with this.

Ref #75