Add a repository test
Also switched from using flags to using an environment variable
(TACO_PORT) for setting the port to run on. This was required because we
now have more than one integration test that depends on the background
workers and the exising method of parsing flags didn't work when you
specify running all tests.